+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.7-preview.4] - 2019-12-09
+
+### Changed
+* Disabled burst for windows/dotnet/collections checks, because it was broken.
+
+## [0.1.7-preview.3] - 2019-12-05
+
+### Changed
+* WindowsBuildTargets will no longer throw trying to print null once a process has terminated.
+
+## [0.1.7-preview.2] - 2019-11-12
+
+### Changed
+* Made fields public on `DotsWindowsTarget` and friends
+
+## [0.1.7-preview] - 2019-10-25
+
+### Changed
+* Updated `com.unity.platforms` to `0.1.7-preview`.
+
 ## [0.1.6-preview] - 2019-10-23
 
 ### Added